Seperated server creation and express definition
Made to ease on tests but better arch anyway
